requirements.txt
streamlit>=1.30.0
python-dotenv>=1.0.0
openai>=1.0.0
tiktoken>=0.5.0
pandas>=2.0.0
python-docx>=1.0.0
PyPDF2>=3.0.0
